decoupling post processor
