+++ libraries/Microsoft.Bot.Builder.Ai.Translation/PostProcessor/CustomDictionaryPostProcessor.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Ai.Translation.PostProcessor
+{
+    class CustomDictionaryPostProcessor
+    {
+    }
+}
