+++ libraries/Microsoft.Bot.Builder.Ai.Translation/PostProcessor/IPostProcessor.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Ai.Translation.PostProcessor
+{
+    public interface IPostProcessor
+    {
+        void Process(out string processedResult);
+    }
+}
