+++ libraries/Microsoft.Bot.Builder.Ai.Translation/PostProcessor/PostProcessorTypesEnum.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Ai.Translation.PostProcessor
+{
+    internal enum PostProcessorTypesEnum
+    {
+        Patterns = 0,
+        Dictionary = 1
+    }
+}
